lock: Introduce dedicated refresh callback
So far our "refresh" callback was the same as our "render" callback.
With this change we split the two. In so doing we can now honor the
Change value reported when ticking the game for "regular" render paths
(i.e., from the "render" callback). This means we ultimately have to
perform fewer render operations. However, we need to make sure to still
force a re-render from the "refresh" callback, so introduce a flag for
that.