Factor out Context type for OpenGL surface management
Our Window type combines logic for system window creation with OpenGL
context/surface setup. As it turns out, though, what we need for the
xlock mode is the means for just initializing OpenGL on an existing
window -- without creating the window ourselves.
As a first step towards this goal, this change factors out the Context
type, which provides the means for initializing an OpenGL context and
managing its state.