Adjust Renderer::on_pre_render() to require Context object
The Renderer::on_pre_render() method holds on to a Window object in an
attempt to make sure that a Window won't vanish while rendering is in
progress.
However, given our desire to decouple the Window management logic from
the OpenGL context management, it actually makes more sense for this
method to require a Context object. In many cases this object will just
be borrowed from the (owned) Window, but in case the window is system
provided only a Context may be available to begin with.
This change adjusts Renderer::on_pre_render() to accept a Context
reference instead of a Window reference.