xlock/mode: Don't create a Window
The xlock framework, by design, takes care of creating the X11 window on
which all the rendering happens. However, so far we have still gone
through the whole process of creating a window ourselves -- if only with
data already provided by xlock.
In winit, window creation has a dependency on an event loop and in the
xlock environment we are not in the business of managing events
ourselves. Version 0.30 of winit furthermore restricts the paths on
which window creation may happen, to the point that we will not be able
to create a window in the xlock environment at all.
Given those two facts, this change decouples our logic from winit and
any window creation business completely. Instead, all we do is make sure
that the OpenGL context is set up properly, so that Tetromino can be
rendered correctly.