Allow game pausing while clearing lines
So far we have simply (and silently) ignored "requests" to pause the
game while lines were being cleared. Conceptually that is a very small
time frame and it should not matter all that much. In reality, it was
mildly irritating to see a key press just being ignored in a fast paced
game.
With this change we make game pausing an unconditional operation: a
"request" will always be satisfied. The main obstacle to doing that in
the past was that if a clearing "animation" was in progress, it wasn't
quite clear how we would convey the end of the animation once the game
is resumed. With this change we solve this issue by just eagerly
transitioning from "clearing" state to "moving", arguing that it is
actually awkward to have a clearing animation continue after a resume
(if one was in progress while pausing).