Add support for dark mode
This change introduces the means for switching between light and dark
mode. In a nutshell, we adjust the clear color as well as the main
background image color to something a little less bright when running in
dark mode. The color mode can be toggled at runtime via a press of F4.