Make the "color mode" a first class entity
Our implementation of color modes has been a big ad hoc, mostly because
it wasn't clear how usable it is. By now the belief is that it is a
useful concept indeed and so with this change we rework the internals a
bit to make this more of a first class concept. Specifically, we
introduce the ColorMode enum which captures the currently active color
mode. We also have the ColorSet type, which comprises both the light as
well as the dark color.