Store Piece color out-of-band
Piece objects really are the one kind of entity that dominate resource
consumption on truly large fields. We already keep their footprint low
by storing the texture in the managing entity as opposed to the Piece
object itself, but we still store the color inside. While not a
particularly big deal, future changes will actually benefit from
introducing an indirection there and it so happens that doing so also
allows us to decrease size of the objects further.
With this change, instead of storing a Color object inside a Piece, we
just store a color index, which is only one fourth the size. We then use
this index to look up the actual color to use when rendering.