Change color mode of more entities
So far the color mode has only affected the background image we use as
well as the overarching clear color. However, some of the piece colors
may look mildly confusing in dark mode, given that they have been
chosen mostly with light mode in mind.
With this change we extend the color mode concept to the walls and
pieces and everything managing them, which includes Stone object as well
as the PreviewStones type. In so doing we allow for changing of colors
of these entities whenever the user toggles the color mode.