Move screen clearing out of Renderer
On some level it makes sense for the Renderer to take care of the
clearing of the screen as part of the "pre-render" phase. In practice,
though, it means that the Renderer now has to know about the clear
color, which unnecessarily convolutes the overall design.
To fix that, this change moves the screen clearing logic out of the
Renderer and into the Game. This way, we basically delegate all color
management to entities inside of the Game object, which helps centralize
color management and control flows when switching color modes.