xlock/mode: Use Tetromino in dark mode
There is an argument to be made that screen locks/screen savers should
be "saving" the monitor and usage of large white surfaces goes against
that credo.
For a lock like ours it probably makes the most sense to use darker
colors and with this change we start up in dark mode instead of light
mode to accomplish exactly that.