xlock/bindings: Set user agent when downloading sources
We currently fail the download of the xlock sources in order to build
bindings, because the first URL we try is for the current release
(5.77), but we attempt to retrieve an older one (5.73). As it turns out,
this URL returns a 406 error as opposed the expected 404 when no user
agent is set. Face -> desk.
Set a user agent when making the request to get the expected behavior.
Also adjust the timeout behavior slightly to be more liberal on slow
Internet connections.