Switch to using winit::application::ApplicationHandler
With winit version 0.30, the crate deprecated the "explicit" event loop
logic that we have been using so far. Instead, everything is intended to
be funneled through an object implementing the ApplicationHandler trait.
This change introduces the internal App type to the program, which
contains the necessary state while implementing the ApplicationHandler
interface. As per the contract, we make sure to create the window only
once we received a "resumed" event. From that point onwards, the
application is able to handle events just as before, but via dedicated
methods instead of a large closure.
Error handling is somewhat convoluted, because there is no way for us to
bubble up an error directly. Rather, we store it inside the App object
itself and once the event loop exits we check if an error was present
and report it accordingly.