Stop returning Change from Game::restart()
It is obvious that restarting the game implies a state change. As such,
it doesn't really make sense to make Game::restart() return a Change
object, because it unconditionally always returns the same value anyway.
This change make the method return nothing, forcing callers to know that
a change occurred.