Disable key repetition for more keys
It doesn't really make sense for the enabling/disabling of the
auto-playing AI to be eligible for key repetition: each press of the
corresponding key should only cause a single state change.
Similarly, it rotating stones using key repetition is just too
uncontrollable. Rather, users are most likely to desire one single per
key press.
This change disables key repetition for the changing of AI as well as
stone rotation.