Start overhaul of KeyState enum
Our current key press/repeat/release logic may occasionally "swallow"
events. What can happen, for example, is that a key is pressed and then
released, even before a "tick" occurred in which we would have reacted
to the press event. This is obviously undesirable behavior.
To fix it, we need to be sure to remember presses unhandled so far, even
if a release was registered already. This change adjusts the KeyState
enum to include a ReleasePending variant and adds a "fire count" to each
of its variants. The fire count basically tracks the number of
undelivered events. Note that this change only introduces this
functionality; it does not yet generate the ReleasePending variant. This
step will come subsequently.