Specify font size in game units
So far we have specified the size of the font we use for rendering the
score and other meta data in pixels. That makes for a nicely relatable
experience with other software that does the same, but it comes with
drawbacks:
- it mixes data from the "physical" layer (the window and, to some
  degree the renderer) with that of the game
- it means that, unless we actively change the font size in response to
  window resizes, we end up with a font size that does not scale with
  the rest of the game components

With this change we instead express the font size in terms of game
units, which alleviates these problems.