Only set viewport on window size change
So far we have set the viewport we use every time we render to the
screen. While conceptually fine, the viewport is ultimately global state
that only changes on certain conditions. We know when it should change,
because we already get informed about window resizes.
Thus, with this change we update the viewport only when we get a window
resize event.