Parametrize Keys::tick() by "change"
Our Keys type provides a way to handle key repetitions. While the core
logic is generic, the kind of "change" being produced as part of its
tick() method is very specific to the rest of the game.
In an attempt to make the logic more general, parametrize Keys::tick()
over the "change" type in use. All we require of it is that it has a
Default implementation and supports bitwise OR'ing of values.