Split up Keys::on_key_event()
So far the Keys::on_key_event() method accepted the winit specific
ElementState enum as an input argument. That arguably is an unnecessary
tight coupling.
With this change we split the method into on_key_press() and
on_key_release(), which encode this state via the function itself,
eliminating this dependency.