Switch over to using our own matrix stacks
With this change we switch over to using our own model view and
perspective matrix stacks (based on xgl::MatrixStack), instead of
relying on OpenGL provides ones (we still indirectly use them, but just
until we switch to using programmable shaders with uniforms instead of
this part of the legacy fixed function pipeline). Our matrix
implementation is provided by the vema crate.