Run some tests in separate process
With upcoming changes we are going to use an actual but empty texture to
represent the "no texture" state. This has trickle down effects in that
we now require a proper OpenGL context for anything that previously used
our dummy invalid texture.
In preparation for that, make sure to run the affected tests in a
separate process and with a proper OpenGL context present.