Remove Texture::invalid() helper method
In the past we used texture ID 0 to represent an "invalid" texture,
which really was meant to act as one not affecting the output. That was
fine in a fixed-function pipeline world, but as we transition towards
using OpenGL 3.3 and programmable shaders, we need to clean this up,
because ID 0 behavior is no longer well specified.
With this change we remove the Texture::invalid() constructor and
instead support creation of an "empty" (white) 1x1 texture in its place.
To not create such a texture over and over every time an ActiveRenderer
is instantiated, move creation into the Renderer object itself.
This has a few trickle down effects:
1) because texture creation is fallible we need to adjust the
   Renderer::new() constructor to be fallible as well
2) because this texture is not bound on ActiveRenderer creation, we have
   to adjust our texture tracking state and the TextureState::Unbound
   variant, specifically, to have an optional still bound texture, not a
   guaranteed one.