Replace Texture with xgl::Texture
Replace our Texture type with that from the xgl crate, to reduce the
number of custom types specific to this program. One of the main
differences is that xgl::Texture is not cloneable or reference counted.
As a result, we now work with Rc<Texture> objects instead of mere
Texture ones.