Replace OpenGL typedefs with Rust builtin types
In the past we used gl::GL* typedefs as a way to stay as close to the
OpenGL API as possible. At the end of the day, doing so provides little
to no benefit and may actively introduce friction.
Replace these typedefs with direct usage of corresponding or more
suitable Rust builtin types to decouple us a little more from more or
less irrelevant OpenGL specifics.