Represent colors in Vertex type as floats
So far we used four bytes to represent a color inside the Vertex type.
That was fine overall and potentially the most suitable representation.
However, in the future we are going to rely on custom programmable
shaders and in GLSL a color is comprised of four floats. There are
probably ways around that, but right now such a micro-optimization is
not the topic of interest and so with this change we use floats to
represent the color information in a vertex.