Don't specify z-coordinate in Vertex type
Our Vertex type contains, among other things, the desired position of
the vertex. Said position is a 3D one, but the game really only is two
dimensional. As a result, we always set the z coordinate to 0.

We ultimately don't really need to provide a zero value, because the
shader can do that for us. Remove the z coordinate from the Vertex type,
shrinking it further and decreasing required memory bandwidth.